refactor: 패키지명 변경 및 position 상수 클래스 추가
