+++ src/main/java/racingcar/constant/PositionConstant.java
@@ -1,0 +1,17 @@
+package racingcar.constant;
+
+public enum PositionConstant {
+    LOW_BOUND(0),
+    HIGH_BOUND(9),
+    FORWARD_THRESHOLD(4);
+
+    private final int value;
+
+    PositionConstant(int value) {
+        this.value = value;
+    }
+
+    public int getValue() {
+        return value;
+    }
+}
