+++ src/main/java/racingcar/constant/SystemMessage.java
@@ -1,4 +1,4 @@
-package racingcar.message;
+package racingcar.constant;
 
 public enum SystemMessage {
     GAME_START_MESSAGE("경주할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)"),
